+++ vite.config.ts
@@ -3,9 +3,15 @@
 
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import tsconfigPaths from 'vite-tsconfig-paths';
 
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react(),
+    tsconfigPaths({
+      projects: ['./tsconfig.app.json'],
+    }),
+  ],
   test: {
     globals: true,
     environment: 'jsdom',
